// src/client/zod.gen.ts
export const zDisciplineRecordCreate = z.object({
    record_date: z.iso.date(),
    type: z.optional(z.union([
        z.string(),
        z.null()
    ])),
    reason: z.optional(z.union([
        z.string(),
        z.null()
    ])),
    display_order: z.optional(z.int()).default(0),
    resume_id: z.optional(z.union([
        z.uuid(),
        z.null()
    ]))
});